refactor: cli input args for meta and consensus pipelines
